Knit different versions of a file from chunk and section options

versions_quarto_multilingual is a function that should be included in the setup chunk of a Quarto document. Its purpose is to write multiple Quarto source files for several versions of a document, specifically when multiple languages are being used for different versions. Python versions must contain the word "python".

versions_quarto_multilingual(
  to_knit = NULL,
  folders = NULL,
  global_eval = TRUE,
  to_jupyter = FALSE
)

Arguments

to_knit

Character vector specifying which versions to write and knit into separate files. If not specified, all versions are produced.

folders

List of versions and subfolder to put them in. Use pattern version_name = folder_name. Default is each version in its own folder.

global_eval

Reset global eval to TRUE? (parent file sets it to FALSE)

to_jupyter

Should a .ipynb document be generated from the python version?

Value

none